+++ tests/testthat/test-stats.R
@@ -805,28 +805,28 @@ test_that("computeMetricMDS works", {
     
     mdsObj <- computeMetricMDS(pwDist, nDim = 2, seed = 0)
     
-    expect_equal(mdsObj$stress, 0.0203635387)
+    expect_equal(stress(mdsObj), 0.0203635387)
     tgtspp <- c(
         11.111867, 13.086625, 4.448069, 37.334795, 34.018645)
     names(tgtspp) <- 1:5
-    expect_equal(mdsObj$spp, tgtspp)
+    expect_equal(spp(mdsObj), tgtspp)
     
     # with no user provided nDim, but (implicit) target pseudo rsquare = 0.95
     mdsObj2 <- computeMetricMDS(pwDist, seed = 0)
     
-    expect_equal(mdsObj$nDim, 2)
-    expect_equal(mdsObj$RSq[2], 0.99843722)
+    expect_equal(nDim(mdsObj2), 2)
+    expect_equal(RSqVec(mdsObj2)[2], 0.99843722)
     
     # with no user provided nDim, but explicit target pseudo rsquare = 0.98
     mdsObj3 <- computeMetricMDS(pwDist, seed = 0, targetPseudoRSq = 0.98)
     
-    expect_equal(mdsObj3$nDim, 2)
-    expect_equal(mdsObj3$RSq[2], 0.99843722)
+    expect_equal(nDim(mdsObj3), 2)
+    expect_equal(RSqVec(mdsObj3)[2], 0.99843722)
     
     # with no user provided nDim, but explicit target pseudo rsquare = 0.999
     mdsObj4 <- computeMetricMDS(pwDist, seed = 0, targetPseudoRSq = 0.999)
     
-    expect_equal(mdsObj4$nDim, 3)
-    expect_equal(mdsObj4$RSq[3], 0.99988906)
+    expect_equal(nDim(mdsObj4), 3)
+    expect_equal(RSqVec(mdsObj4)[3], 0.99988906)
     
 })